requirements.txt
streamlit
cerebras-cloud-sdk
PyPDF2
python-docx
python-dotenv